More 'MoreLikeRefactoring'-Component refactoring
