+++ tests/Solarium/Tests/QueryType/Select/Query/QueryTest.php
@@ -31,7 +31,7 @@
 
 namespace Solarium\Tests\QueryType\Select\Query;
 
-use Solarium\Component\MoreLikeThis\Select\Query\SelectQueryComponent;
+use Solarium\Component\MoreLikeThis\Select\Query as SelectQuery;
 use Solarium\Core\Client\Client;
 use Solarium\QueryType\Select\Query\FilterQuery;
 use Solarium\QueryType\Select\Query\Query;
@@ -491,7 +491,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
 
     public function testSetAndGetComponents()
     {
-        $mlt = new SelectQueryComponent();
+        $mlt = new SelectQuery();
         $this->query->setComponent('mlt', $mlt);
 
         $this->assertEquals(
@@ -502,7 +502,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
 
     public function testSetAndGetComponent()
     {
-        $mlt = new SelectQueryComponent();
+        $mlt = new SelectQuery();
         $this->query->setComponent('mlt', $mlt);
 
         $this->assertEquals(
@@ -513,7 +513,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
 
     public function testSetAndGetComponentQueryInstance()
     {
-        $mlt = new SelectQueryComponent();
+        $mlt = new SelectQuery();
         $this->query->setComponent('mlt', $mlt);
 
         $this->assertEquals(
@@ -538,7 +538,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
 
     public function testRemoveComponent()
     {
-        $mlt = new SelectQueryComponent();
+        $mlt = new SelectQuery();
         $this->query->setComponent('mlt', $mlt);
 
         $this->assertEquals(
@@ -556,7 +556,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
 
     public function testRemoveComponentWithObjectInput()
     {
-        $mlt = new SelectQueryComponent();
+        $mlt = new SelectQuery();
         $this->query->setComponent('mlt', $mlt);
 
         $this->assertEquals(
@@ -577,7 +577,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
         $mlt = $this->query->getMoreLikeThis();
 
         $this->assertEquals(
-            'Solarium\Component\MoreLikeThis\Select\Query\SelectQueryComponent',
+            'Solarium\Component\MoreLikeThis\Select\Query',
             get_class($mlt)
         );
     }
